requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
Flask-Caching==2.1.0
Flask-Limiter==3.5.0
Flask-SQLAlchemy==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
limits==3.7.0
markdown==3.5
MarkupSafe==3.0.2
ordered-set==4.1.0
PyMuPDF==1.25.5
python-dotenv==1.1.0
requests==2.32.3
SQLAlchemy==2.0.29
typing_extensions==4.9.0
urllib3==2.4.0
Werkzeug==3.1.0
